requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
fastapi==0.121.3
h11==0.16.0
idna==3.11
pydantic==2.12.4
pydantic_core==2.41.5
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
gspread==6.1.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pymysql==1.1.1
httpx==0.27.2
openai==1.54.3
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
Pillow==10.4.0
python-multipart==0.0.20
